fix: focus issue on translate modal, remove duplicate code
