+++ src/components/Modal/Translate.vue
@@ -20,7 +20,7 @@
   -->
 
 <template>
-	<NcModal size="large" @close="$emit('close')">
+	<NcModal :show="show" size="large" @close="$emit('close')">
 		<div class="translate-dialog">
 			<h2>{{ t('text', 'Translate') }}</h2>
 			<em>{{ t('text', 'To translate individual parts of the text, select it before using the translate function.') }}</em>
@@ -106,6 +106,10 @@ export default {
 		useIsMobileMixin,
 	],
 	props: {
+		show: {
+			type: Boolean,
+			default: false,
+		},
 		content: {
 			type: String,
 			default: '',
@@ -165,6 +169,9 @@ export default {
 		},
 	},
 	watch: {
+		content() {
+			this.input = this.content
+		},
 		input() {
 			this.result = null
 			this.error = null
@@ -246,6 +253,7 @@ export default {
 		resize: none;
 		box-sizing: border-box;
 		overflow-y: auto;
+		min-height: 62px;
 		max-height: 58vh;
 	}
 }
